+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v3.0.2
+
+* `lib/zypper`: POSIX ready, tests added.
+
 ## v3.0.1
 
 * More `POSIX` ports: `pkg_tools`, `conda`, `tlmgr`, `dnf`, `swupd`, `yum`, `macports`